Demonstration of issue https://github.com/apollographql/apollo-client-devtools/issues/477
In Apollo dev tools, notice that the result of executing the following query includes
people id, even though it was not asked for

```
query AllPeople {
  people @client {
    name
  }
}
```

However, put a breakpoint on line 19 of App.js and notice that the apollo client
itself is correctly filtering out data not requested before returning it to
the caller.

Apollo dev tools should do the same in its GraphiQL tab.